+++ ui/src/components/navbar.vue
@@ -1,8 +1,8 @@
 <template lang="pug">
 .tw-relative(style="z-index: 99")
   .top-navi-bar-minified(v-if="$route.meta.config.showMinifiedNavbar")
-    v-btn.text-default.included(@click="toggleNavi" icon height="38px" width="38px")
-      v-icon.text-transparent {{ showSidebar ? icons['mdiArrowLeftThick'] : icons['mdiArrowRightThick'] }}
+    v-btn.text-muted.included(@click="toggleNavi" fab elevation="1" height="38px" width="38px" color="rgba(0, 0, 0, 0.5)" retain-focus-on-click)
+      v-icon {{ showSidebar ? icons['mdiArrowLeftThick'] : icons['mdiArrowRightThick'] }}
   v-app-bar.top-navi-bar.pt-safe(v-else height="64px" :class="($route.meta.config.fixedNavbar ? 'top-navi-bar-fixed ' : '') + (extendSidebar ? 'extended-sidebar' : '')")
     .navi-wrap.pl-safe.pr-safe
       v-btn.text-default.included(@click="toggleNavi" icon height="38px" width="38px")
@@ -137,7 +137,7 @@ span >>> .v-badge__badge::after {
 }
 
 .text-transparent {
-  color: rgba(255, 255, 255, 0.3) !important;
+  color: rgba(255, 255, 255, 0.6) !important;
 }
 
 .badge-text {